+++ Common/src/main/java/net/ixdarklord/packmger/compat/CurseAPI.java
@@ -3,8 +3,6 @@ package net.ixdarklord.packmger.compat;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 import java.util.*;