+++ Common/src/main/java/net/ixdarklord/packmger/compat/fancymenu/ModpackUpdateRequirement.java
@@ -1,23 +1,22 @@
 package net.ixdarklord.packmger.compat.fancymenu;
 
-import de.keksuccino.fancymenu.api.visibilityrequirements.VisibilityRequirement;
-import de.keksuccino.konkrete.input.CharacterFilter;
+import de.keksuccino.fancymenu.menu.fancy.helper.ui.texteditor.TextEditorFormattingRule;
+import de.keksuccino.fancymenu.menu.loadingrequirement.v2.LoadingRequirement;
 import net.ixdarklord.packmger.client.handler.WindowHandler;
 import net.ixdarklord.packmger.compat.CurseAPI;
 import net.ixdarklord.packmger.config.ConfigHandler;
-import net.ixdarklord.packmger.config.ConfigHandler.CLIENT.KeyData;
-import net.ixdarklord.packmger.core.Constants;
 import net.ixdarklord.packmger.util.ManagerUtils;
 import net.ixdarklord.packmger.util.VersionUtils;
 import net.ixdarklord.packmger.util.WebUtils;
 import net.minecraft.SharedConstants;
+import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import java.util.*;
 
 import static net.ixdarklord.packmger.compat.fancymenu.ModpackCheckingUpdateButton.isUpdateAvailable;
 
-public class ModpackVersionRequirement extends VisibilityRequirement {
+public class ModpackUpdateRequirement extends LoadingRequirement {
 
     private static final String IDENTIFIER = ConfigHandler.CLIENT.MODPACK_UPDATE_IDENTIFIER.get();
     private static final String CURRENT_VERSION = ConfigHandler.CLIENT.MODPACK_VERSION.get();
@@ -25,27 +24,15 @@ public class ModpackVersionRequirement extends VisibilityRequirement {
     private static final String REQUIREMENT_NAME = "Is Update Available";
     private static final String[] REQUIREMENT_DESC = new String[] {
             "This requirement will behave depending on new update availability!",
-            "If there is an update, it will sit true. Otherwise, it will sit to false."
+            "If there is an update, it will sit to true. Otherwise, it will sit to false."
     };
-    private static Map<String, List<String>> cachedValues = new HashMap<>();
-    private static List<String> invalidURL = new ArrayList<>();
-    private static List<String> currentlyUpdatingPlaceholders = new ArrayList<>();
+    private static final Map<String, List<String>> cachedValues = new HashMap<>();
+    private static final List<String> invalidURL = new ArrayList<>();
+    private static final List<String> currentlyUpdatingPlaceholders = new ArrayList<>();
     VersionUtils VC = new VersionUtils();
 
-    public ModpackVersionRequirement() {
-        super("modpack_version_requirement");
-    }
-
-    public static void reloadMenu(boolean isDebugMode) {
-        try {
-            WindowHandler.CACHED_NEW_VERSION = null;
-            isUpdateAvailable = false;
-            cachedValues.clear();
-            invalidURL.clear();
-            if (isDebugMode) Constants.LOGGER.info("ModpackVersionRequirement cache successfully cleared!");
-        } catch (Exception ex) {
-            ex.printStackTrace();
-        }
+    public ModpackUpdateRequirement() {
+        super("modpack_update_requirement");
     }
 
     @Override
@@ -54,7 +41,7 @@ public class ModpackVersionRequirement extends VisibilityRequirement {
             if (!isURLInvalid(UPDATE_KEY)) {
                 List<String> version = cachedValues.get("version");
                 if (version != null) {
-                    String ID = cachedValues.get(KeyData.IDENTIFIER.ID).toString();
+                    String ID = cachedValues.get(ConfigHandler.CLIENT.KeyData.IDENTIFIER.ID).toString();
                     ID = ID.substring(1, ID.length()-1);
                     if (ID.equals(IDENTIFIER)) {
                         String NEW_VERSION = cachedValues.get("version").toString().replaceAll("[\\(\\)\\[\\]\\{\\}]", "");
@@ -72,7 +59,7 @@ public class ModpackVersionRequirement extends VisibilityRequirement {
                             ModpackCheckingUpdateButton.setInvalidManifest(cachedValues);
                     }
                 } else {
-                    cacheValues(KeyData.IDENTIFIER.ID, "version", UPDATE_KEY, "modpack.identifier", "modpack.version");
+                    cacheValues(ConfigHandler.CLIENT.KeyData.IDENTIFIER.ID, "version", UPDATE_KEY, "modpack.identifier", "modpack.version");
                 }
             } else {
                 if (!ModpackCheckingUpdateButton.isConnectionFailed)
@@ -129,13 +116,18 @@ public class ModpackVersionRequirement extends VisibilityRequirement {
     }
 
     @Override
-    public String getDisplayName() {
+    public @NotNull String getDisplayName() {
         return REQUIREMENT_NAME;
     }
 
     @Override
     public List<String> getDescription() {
         return List.of(REQUIREMENT_DESC);
+    }
+
+    @Override
+    public @Nullable String getCategory() {
+        return null;
     }
 
     @Override
@@ -154,7 +146,7 @@ public class ModpackVersionRequirement extends VisibilityRequirement {
     }
 
     @Override
-    public CharacterFilter getValueInputFieldFilter() {
+    public @Nullable List<TextEditorFormattingRule> getValueFormattingRules() {
         return null;
     }
 }