+++ Fabric/src/main/java/net/ixdarklord/packmger/core/FabricClientSetup.java
@@ -6,6 +6,7 @@ import net.ixdarklord.packmger.config.ConfigHandler;
 import net.ixdarklord.packmger.event.ClientEvents;
 
 public class FabricClientSetup implements ClientModInitializer {
+
     @Override
     public void onInitializeClient() {
         ConfigHandler.initializeFiles();