+++ Forge/src/main/java/net/ixdarklord/packmger/mixin/client/MixinBrandingControl.java
@@ -8,7 +8,6 @@ import net.minecraftforge.internal.BrandingControl;
 import net.minecraftforge.versions.forge.ForgeVersion;
 import net.minecraftforge.versions.mcp.MCPVersion;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Shadow;
 
 import java.util.List;
 
@@ -22,9 +21,11 @@ public abstract class MixinBrandingControl {
             ImmutableList.Builder<String> brd = ImmutableList.builder();
             brd.add("Forge " + ForgeVersion.getVersion());
             brd.add("Minecraft " + MCPVersion.getMCVersion());
+
             if (!ConfigHandler.CLIENT.getTitleName().equalsIgnoreCase("Minecraft"))
                 brd.add(String.format("%s %s", ConfigHandler.CLIENT.getTitleName(), ConfigHandler.CLIENT.MODPACK_VERSION.get()));
             brd.add("MCP " + MCPVersion.getMCPVersion());
+
             int tModCount = ModList.get().size();
             brd.add(ForgeI18n.parseMessage("fml.menu.loadingmods", tModCount));
             brandings = brd.build();